refactor: 디바이스 등록 모달·연결 대기 로직 공용화 (#101)
온보딩에서도 기기 등록과 연결 대기가 필요해 설정 페이지 전용이던 코드를 공유한다.
동작 변화는 없다.

- DeviceRegisterModal을 shared/components로 이동
  (온보딩이 pages/setting을 참조하는 역참조를 만들지 않기 위함)
- useDeviceSection의 대기/타임아웃 로직을 useDeviceConnectionWait 훅으로 추출
  useDeviceSection은 모달 3개에 삭제·이름변경까지 들고 있어 통째로는 재사용할 수 없다.
- DEVICE_CONNECT_TIMEOUT_MS를 shared/constants로 이동
  (shared 훅이 pages를 참조하지 않도록)

새 훅은 조회를 하지 않고 isConnected만 인자로 받는다. 폴링 주기가 화면마다 다르기 때문이다
(온보딩은 즉각 반응이 필요해 1초, 설정은 상태 표시라 5초).
