+++ src/shared/components/DeviceRegisterModal.tsx
